Refactored into multiple packages
