+++ pkg/patch/filestore/filestore.go
@@ -1,0 +1,32 @@
+// SPDX-FileCopyrightText: 2024 Risk.Ident GmbH <contact@riskident.com>
+//
+// SPDX-License-Identifier: GPL-3.0-or-later
+//
+// This program is free software: you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by the
+// Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+// FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+// more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+// Package filestore contains a simple and minimal abstraction over
+// OS operations for reading and writing to files.
+package filestore
+
+// FileStore is a minimal abstraction over reading and writing files
+// that is used in the patching steps (e.g yaml patch, regex patch, etc).
+//
+// This abstraction allows us to mock the filesystem during testing,
+// as well as adding additional features like caching when you have
+// multiple patches on the same file (see [Cached]).
+type FileStore interface {
+	ReadFile(path string) ([]byte, error)
+	WriteFile(path string, content []byte) error
+	Close() error
+}
