+++ pkg/server/config_trypackage.go
@@ -25,7 +25,6 @@ import (
 	"text/template"
 
 	"github.com/RiskIdent/jelease/pkg/config"
-	"github.com/RiskIdent/jelease/pkg/patch"
 	"github.com/RiskIdent/jelease/templates/pages"
 	"github.com/gin-gonic/gin"
 	"github.com/rs/zerolog/log"
@@ -102,7 +101,7 @@ func (s HTTPServer) handlePostConfigTryPackage(c *gin.Context) {
 	cfgClone.DryRun = true
 	patcherClone := s.patcher.CloneWithConfig(&cfgClone)
 
-	tmplCtx := patch.TemplateContext{
+	tmplCtx := config.TemplateContext{
 		Package: model.Package.Name,
 		Version: model.Version,
 	}
